add mvc project and custom token

+++ SmartBill/SmartBill.Entities/Helpers/JWT.cs
@@ -11,7 +11,7 @@ namespace SmartBill.Entities.Helpers
         public string Key { get; set; }
         public string Issuer { get; set; }
         public string Audience { get; set; }
-        public int DurationInDays { get; set; }
+        public double DurationInDays { get; set; }
     }
 }
